refactor(runtime): extract turn/stream_consume.rs into agent/turn — #7415
